Properly support EGL on top of DirectFB.
This enables us to remove a workaround when calling
eglCreateWindowSurface() that used to ignore the native window created
by the Ozone surface and passed 0 instead. That works with ST SDK2 EGL
on DirectFB but breaks other platforms such as Nexus that really need
the native window.

Note that the support for the Hisilicon platform has been removed as we
don't use it and I have no board to test it anyway.